+++ sdk/src/driftClient.ts
@@ -181,6 +181,8 @@ import { gprcDriftClientAccountSubscriber } from './accounts/grpcDriftClientAcco
 import nacl from 'tweetnacl';
 import { Slothash } from './slot/SlothashSubscriber';
 import { getOracleId } from './oracles/oracleId';
+import { SignedSwiftOrderParams } from './swift/types';
+import { sha256 } from '@noble/hashes/sha256';
 
 type RemainingAccountParams = {
 	userAccounts: UserAccount[];
@@ -1045,17 +1047,14 @@ export class DriftClient {
 	}
 
 	public async initializeSwiftUserOrders(
-		userAccountPublicKey: PublicKey,
+		authority: PublicKey,
 		numOrders: number,
 		txParams?: TxParams
 	): Promise<[TransactionSignature, PublicKey]> {
 		const initializeIxs = [];
 
 		const [swiftUserAccountPublicKey, initializeUserAccountIx] =
-			await this.getInitializeSwiftUserOrdersAccountIx(
-				userAccountPublicKey,
-				numOrders
-			);
+			await this.getInitializeSwiftUserOrdersAccountIx(authority, numOrders);
 		initializeIxs.push(initializeUserAccountIx);
 		const tx = await this.buildTransaction(initializeIxs, txParams);
 		const { txSig } = await this.sendTransaction(tx, [], this.opts);
@@ -1064,19 +1063,18 @@ export class DriftClient {
 	}
 
 	async getInitializeSwiftUserOrdersAccountIx(
-		userAccountPublicKey: PublicKey,
+		authority: PublicKey,
 		numOrders: number
 	): Promise<[PublicKey, TransactionInstruction]> {
 		const swiftUserAccountPublicKey = getSwiftUserAccountPublicKey(
 			this.program.programId,
-			userAccountPublicKey
+			authority
 		);
 		const initializeUserAccountIx =
 			await this.program.instruction.initializeSwiftUserOrders(numOrders, {
 				accounts: {
 					swiftUserOrders: swiftUserAccountPublicKey,
 					authority: this.wallet.publicKey,
-					user: userAccountPublicKey,
 					payer: this.wallet.publicKey,
 					rent: anchor.web3.SYSVAR_RENT_PUBKEY,
 					systemProgram: anchor.web3.SystemProgram.programId,
@@ -1087,12 +1085,12 @@ export class DriftClient {
 	}
 
 	public async resizeSwiftUserOrders(
-		userAccountPublicKey: PublicKey,
+		authority: PublicKey,
 		numOrders: number,
 		txParams?: TxParams
 	): Promise<TransactionSignature> {
 		const resizeUserAccountIx = await this.getResizeSwiftUserOrdersInstruction(
-			userAccountPublicKey,
+			authority,
 			numOrders
 		);
 		const tx = await this.buildTransaction([resizeUserAccountIx], txParams);
@@ -1102,19 +1100,18 @@ export class DriftClient {
 	}
 
 	async getResizeSwiftUserOrdersInstruction(
-		userAccountPublicKey: PublicKey,
+		authority: PublicKey,
 		numOrders: number
 	): Promise<TransactionInstruction> {
 		const swiftUserAccountPublicKey = getSwiftUserAccountPublicKey(
 			this.program.programId,
-			userAccountPublicKey
+			authority
 		);
 		const resizeUserAccountIx =
 			await this.program.instruction.resizeSwiftUserOrders(numOrders, {
 				accounts: {
 					swiftUserOrders: swiftUserAccountPublicKey,
 					authority: this.wallet.publicKey,
-					user: userAccountPublicKey,
 					systemProgram: anchor.web3.SystemProgram.programId,
 				},
 			});
@@ -1712,16 +1709,9 @@ export class DriftClient {
 	}
 
 	public async deleteSwiftUserOrders(
-		subAccountId = 0,
 		txParams?: TxParams
 	): Promise<TransactionSignature> {
-		const userAccountPublicKey = getUserAccountPublicKeySync(
-			this.program.programId,
-			this.wallet.publicKey,
-			subAccountId
-		);
-
-		const ix = await this.getSwiftUserOrdersDeletionIx(userAccountPublicKey);
+		const ix = await this.getSwiftUserOrdersDeletionIx(this.wallet.publicKey);
 
 		const { txSig } = await this.sendTransaction(
 			await this.buildTransaction(ix, txParams),
@@ -1729,20 +1719,16 @@ export class DriftClient {
 			this.opts
 		);
 
-		const userMapKey = this.getUserMapKey(subAccountId, this.wallet.publicKey);
-		await this.users.get(userMapKey)?.unsubscribe();
-		this.users.delete(userMapKey);
-
 		return txSig;
 	}
 
-	public async getSwiftUserOrdersDeletionIx(userAccountPublicKey: PublicKey) {
+	public async getSwiftUserOrdersDeletionIx(authority: PublicKey) {
 		const ix = await this.program.instruction.deleteSwiftUserOrders({
 			accounts: {
-				user: userAccountPublicKey,
+				user: authority,
 				swiftUserOrders: getSwiftUserAccountPublicKey(
 					this.program.programId,
-					userAccountPublicKey
+					authority
 				),
 				authority: this.wallet.publicKey,
 				state: await this.getStatePublicKey(),
@@ -5882,27 +5868,42 @@ export class DriftClient {
 
 	public signSwiftOrderParamsMessage(
 		orderParamsMessage: SwiftOrderParamsMessage
-	): Buffer {
-		const takerOrderParamsMessage =
-			this.encodeSwiftOrderParamsMessage(orderParamsMessage);
-		return this.signMessage(takerOrderParamsMessage);
+	): SignedSwiftOrderParams {
+		const borshBuf = this.encodeSwiftOrderParamsMessage(orderParamsMessage);
+		const orderParams = Buffer.from(borshBuf.toString('hex'));
+		return {
+			orderParams,
+			signature: this.signMessage(Buffer.from(borshBuf.toString('hex'))),
+		};
 	}
 
+	/*
+	 * Borsh encode swift taker order params
+	 */
 	public encodeSwiftOrderParamsMessage(
 		orderParamsMessage: SwiftOrderParamsMessage
 	): Buffer {
-		return this.program.coder.types.encode(
-			'SwiftOrderParamsMessage',
-			orderParamsMessage
-		);
+		const anchorIxName = 'global' + ':' + 'swiftOrderMessageParams';
+		const prefix = Buffer.from(sha256(anchorIxName).slice(0, 8));
+		const buf = Buffer.concat([
+			prefix,
+			this.program.coder.types.encode(
+				'SwiftOrderParamsMessage',
+				orderParamsMessage
+			),
+		]);
+		return buf;
 	}
 
+	/*
+	 * Decode swift taker order params from borsh buffer
+	 */
 	public decodeSwiftOrderParamsMessage(
 		encodedMessage: Buffer
 	): SwiftOrderParamsMessage {
 		return this.program.coder.types.decode(
 			'SwiftOrderParamsMessage',
-			encodedMessage
+			encodedMessage.slice(8) // assumes discriminator
 		);
 	}
 
@@ -5914,8 +5915,7 @@ export class DriftClient {
 	}
 
 	public async placeSwiftTakerOrder(
-		swiftOrderParamsMessage: Buffer,
-		swiftOrderParamsSignature: Buffer,
+		signedSwiftOrderParams: SignedSwiftOrderParams,
 		marketIndex: number,
 		takerInfo: {
 			taker: PublicKey;
@@ -5927,8 +5927,7 @@ export class DriftClient {
 		txParams?: TxParams
 	): Promise<TransactionSignature> {
 		const ixs = await this.getPlaceSwiftTakerPerpOrderIxs(
-			swiftOrderParamsMessage,
-			swiftOrderParamsSignature,
+			signedSwiftOrderParams,
 			marketIndex,
 			takerInfo,
 			undefined,
@@ -5944,8 +5943,7 @@ export class DriftClient {
 	}
 
 	public async getPlaceSwiftTakerPerpOrderIxs(
-		encodedSwiftOrderParamsMessage: Buffer,
-		swiftOrderParamsSignature: Buffer,
+		signedSwiftOrderParams: SignedSwiftOrderParams,
 		marketIndex: number,
 		takerInfo: {
 			taker: PublicKey;
@@ -5969,13 +5967,15 @@ export class DriftClient {
 		const authorityToUse = authority || takerInfo.takerUserAccount.authority;
 
 		const messageLengthBuffer = Buffer.alloc(2);
-		messageLengthBuffer.writeUInt16LE(encodedSwiftOrderParamsMessage.length);
+		messageLengthBuffer.writeUInt16LE(
+			signedSwiftOrderParams.orderParams.length
+		);
 
 		const swiftIxData = Buffer.concat([
-			swiftOrderParamsSignature,
+			signedSwiftOrderParams.signature,
 			authorityToUse.toBytes(),
 			messageLengthBuffer,
-			encodedSwiftOrderParamsMessage,
+			signedSwiftOrderParams.orderParams,
 		]);
 
 		const swiftOrderParamsSignatureIx = createMinimalEd25519VerifyIx(
@@ -5993,7 +5993,7 @@ export class DriftClient {
 					userStats: takerInfo.takerStats,
 					swiftUserOrders: getSwiftUserAccountPublicKey(
 						this.program.programId,
-						takerInfo.taker
+						takerInfo.takerUserAccount.authority
 					),
 					authority: this.wallet.publicKey,
 					ixSysvar: SYSVAR_INSTRUCTIONS_PUBKEY,
@@ -6005,8 +6005,7 @@ export class DriftClient {
 	}
 
 	public async placeAndMakeSwiftPerpOrder(
-		encodedSwiftOrderParamsMessage: Buffer,
-		swiftOrderParamsSignature: Buffer,
+		signedSwiftOrderParams: SignedSwiftOrderParams,
 		swiftOrderUuid: Uint8Array,
 		takerInfo: {
 			taker: PublicKey;
@@ -6021,8 +6020,7 @@ export class DriftClient {
 		overrideIxCount?: number
 	): Promise<TransactionSignature> {
 		const ixs = await this.getPlaceAndMakeSwiftPerpOrderIxs(
-			encodedSwiftOrderParamsMessage,
-			swiftOrderParamsSignature,
+			signedSwiftOrderParams,
 			swiftOrderUuid,
 			takerInfo,
 			orderParams,
@@ -6042,8 +6040,7 @@ export class DriftClient {
 	}
 
 	public async getPlaceAndMakeSwiftPerpOrderIxs(
-		encodedSwiftOrderParamsMessage: Buffer,
-		swiftOrderParamsSignature: Buffer,
+		signedSwiftOrderParams: SignedSwiftOrderParams,
 		swiftOrderUuid: Uint8Array,
 		takerInfo: {
 			taker: PublicKey;
@@ -6058,8 +6055,7 @@ export class DriftClient {
 	): Promise<TransactionInstruction[]> {
 		const [swiftOrderSignatureIx, placeTakerSwiftPerpOrderIx] =
 			await this.getPlaceSwiftTakerPerpOrderIxs(
-				encodedSwiftOrderParamsMessage,
-				swiftOrderParamsSignature,
+				signedSwiftOrderParams,
 				orderParams.marketIndex,
 				takerInfo,
 				undefined,
@@ -6107,7 +6103,7 @@ export class DriftClient {
 						authority: this.wallet.publicKey,
 						takerSwiftUserOrders: getSwiftUserAccountPublicKey(
 							this.program.programId,
-							takerInfo.taker
+							takerInfo.takerUserAccount.authority
 						),
 					},
 					remainingAccounts,
@@ -7215,6 +7211,274 @@ export class DriftClient {
 				remainingAccounts: remainingAccounts,
 			}
 		);
+	}
+
+	public async getJupiterLiquidateSpotWithSwapIxV6({
+		jupiterClient,
+		liabilityMarketIndex,
+		assetMarketIndex,
+		swapAmount,
+		assetTokenAccount,
+		liabilityTokenAccount,
+		slippageBps,
+		swapMode,
+		onlyDirectRoutes,
+		quote,
+		userAccount,
+		userAccountPublicKey,
+		userStatsAccountPublicKey,
+		liquidatorSubAccountId,
+		maxAccounts,
+	}: {
+		jupiterClient: JupiterClient;
+		liabilityMarketIndex: number;
+		assetMarketIndex: number;
+		swapAmount: BN;
+		assetTokenAccount?: PublicKey;
+		liabilityTokenAccount?: PublicKey;
+		slippageBps?: number;
+		swapMode?: SwapMode;
+		onlyDirectRoutes?: boolean;
+		quote?: QuoteResponse;
+		userAccount: UserAccount;
+		userAccountPublicKey: PublicKey;
+		userStatsAccountPublicKey: PublicKey;
+		liquidatorSubAccountId?: number;
+		maxAccounts?: number;
+	}): Promise<{
+		ixs: TransactionInstruction[];
+		lookupTables: AddressLookupTableAccount[];
+	}> {
+		const liabilityMarket = this.getSpotMarketAccount(liabilityMarketIndex);
+		const assetMarket = this.getSpotMarketAccount(assetMarketIndex);
+
+		if (!quote) {
+			const fetchedQuote = await jupiterClient.getQuote({
+				inputMint: assetMarket.mint,
+				outputMint: liabilityMarket.mint,
+				amount: swapAmount,
+				slippageBps,
+				swapMode,
+				onlyDirectRoutes,
+				maxAccounts,
+			});
+
+			quote = fetchedQuote;
+		}
+
+		if (!quote) {
+			throw new Error("Could not fetch Jupiter's quote. Please try again.");
+		}
+
+		const amountIn = new BN(quote.inAmount);
+
+		const transaction = await jupiterClient.getSwap({
+			quote,
+			userPublicKey: this.provider.wallet.publicKey,
+			slippageBps,
+		});
+
+		const { transactionMessage, lookupTables } =
+			await jupiterClient.getTransactionMessageAndLookupTables({
+				transaction,
+			});
+
+		const jupiterInstructions = jupiterClient.getJupiterInstructions({
+			transactionMessage,
+			inputMint: assetMarket.mint,
+			outputMint: liabilityMarket.mint,
+		});
+
+		const preInstructions = [];
+		if (!liabilityTokenAccount) {
+			const tokenProgram = this.getTokenProgramForSpotMarket(liabilityMarket);
+			liabilityTokenAccount = await this.getAssociatedTokenAccount(
+				liabilityMarket.marketIndex,
+				false,
+				tokenProgram
+			);
+
+			preInstructions.push(
+				this.createAssociatedTokenAccountIdempotentInstruction(
+					liabilityTokenAccount,
+					this.provider.wallet.publicKey,
+					this.provider.wallet.publicKey,
+					liabilityMarket.mint,
+					tokenProgram
+				)
+			);
+		}
+
+		if (!assetTokenAccount) {
+			const tokenProgram = this.getTokenProgramForSpotMarket(assetMarket);
+			assetTokenAccount = await this.getAssociatedTokenAccount(
+				assetMarket.marketIndex,
+				false,
+				tokenProgram
+			);
+
+			preInstructions.push(
+				this.createAssociatedTokenAccountIdempotentInstruction(
+					assetTokenAccount,
+					this.provider.wallet.publicKey,
+					this.provider.wallet.publicKey,
+					assetMarket.mint,
+					tokenProgram
+				)
+			);
+		}
+
+		const { beginSwapIx, endSwapIx } = await this.getLiquidateSpotWithSwapIx({
+			liabilityMarketIndex,
+			assetMarketIndex,
+			swapAmount: amountIn,
+			assetTokenAccount,
+			liabilityTokenAccount,
+			userAccount,
+			userAccountPublicKey,
+			userStatsAccountPublicKey,
+			liquidatorSubAccountId,
+		});
+
+		const ixs = [
+			...preInstructions,
+			beginSwapIx,
+			...jupiterInstructions,
+			endSwapIx,
+		];
+
+		return { ixs, lookupTables };
+	}
+
+	/**
+	 * Get the drift liquidate_spot_with_swap instructions
+	 *
+	 * @param liabilityMarketIndex the market index of the token you're buying
+	 * @param assetMarketIndex the market index of the token you're selling
+	 * @param amountIn the amount of the token to sell
+	 * @param assetTokenAccount the token account to move the tokens being sold
+	 * @param liabilityTokenAccount the token account to receive the tokens being bought
+	 * @param userAccount
+	 * @param userAccountPublicKey
+	 * @param userStatsAccountPublicKey
+	 */
+	public async getLiquidateSpotWithSwapIx({
+		liabilityMarketIndex,
+		assetMarketIndex,
+		swapAmount: swapAmount,
+		assetTokenAccount,
+		liabilityTokenAccount,
+		userAccount,
+		userAccountPublicKey,
+		userStatsAccountPublicKey,
+		liquidatorSubAccountId,
+	}: {
+		liabilityMarketIndex: number;
+		assetMarketIndex: number;
+		swapAmount: BN;
+		assetTokenAccount: PublicKey;
+		liabilityTokenAccount: PublicKey;
+		userAccount: UserAccount;
+		userAccountPublicKey: PublicKey;
+		userStatsAccountPublicKey: PublicKey;
+		liquidatorSubAccountId?: number;
+	}): Promise<{
+		beginSwapIx: TransactionInstruction;
+		endSwapIx: TransactionInstruction;
+	}> {
+		const liquidatorAccountPublicKey = await this.getUserAccountPublicKey(
+			liquidatorSubAccountId
+		);
+		const liquidatorStatsPublicKey = this.getUserStatsAccountPublicKey();
+
+		const userAccounts = [userAccount];
+		const remainingAccounts = this.getRemainingAccounts({
+			userAccounts,
+			writableSpotMarketIndexes: [liabilityMarketIndex, assetMarketIndex],
+			readableSpotMarketIndexes: [QUOTE_SPOT_MARKET_INDEX],
+		});
+
+		const liabilitySpotMarket = this.getSpotMarketAccount(liabilityMarketIndex);
+		const assetSpotMarket = this.getSpotMarketAccount(assetMarketIndex);
+
+		const liabilityTokenProgram =
+			this.getTokenProgramForSpotMarket(liabilitySpotMarket);
+		const assetTokenProgram =
+			this.getTokenProgramForSpotMarket(assetSpotMarket);
+
+		if (!liabilityTokenProgram.equals(assetTokenProgram)) {
+			remainingAccounts.push({
+				pubkey: liabilityTokenProgram,
+				isWritable: false,
+				isSigner: false,
+			});
+		}
+
+		if (
+			liabilitySpotMarket.tokenProgram === 1 ||
+			assetSpotMarket.tokenProgram === 1
+		) {
+			remainingAccounts.push({
+				pubkey: assetSpotMarket.mint,
+				isWritable: false,
+				isSigner: false,
+			});
+			remainingAccounts.push({
+				pubkey: liabilitySpotMarket.mint,
+				isWritable: false,
+				isSigner: false,
+			});
+		}
+
+		const beginSwapIx =
+			await this.program.instruction.liquidateSpotWithSwapBegin(
+				assetMarketIndex,
+				liabilityMarketIndex,
+				swapAmount,
+				{
+					accounts: {
+						state: await this.getStatePublicKey(),
+						user: userAccountPublicKey,
+						userStats: userStatsAccountPublicKey,
+						liquidator: liquidatorAccountPublicKey,
+						liquidatorStats: liquidatorStatsPublicKey,
+						authority: this.wallet.publicKey,
+						liabilitySpotMarketVault: liabilitySpotMarket.vault,
+						assetSpotMarketVault: assetSpotMarket.vault,
+						assetTokenAccount: assetTokenAccount,
+						liabilityTokenAccount: liabilityTokenAccount,
+						tokenProgram: assetTokenProgram,
+						driftSigner: this.getStateAccount().signer,
+						instructions: anchor.web3.SYSVAR_INSTRUCTIONS_PUBKEY,
+					},
+					remainingAccounts,
+				}
+			);
+
+		const endSwapIx = await this.program.instruction.liquidateSpotWithSwapEnd(
+			assetMarketIndex,
+			liabilityMarketIndex,
+			{
+				accounts: {
+					state: await this.getStatePublicKey(),
+					user: userAccountPublicKey,
+					userStats: userStatsAccountPublicKey,
+					liquidator: liquidatorAccountPublicKey,
+					liquidatorStats: liquidatorStatsPublicKey,
+					authority: this.wallet.publicKey,
+					liabilitySpotMarketVault: liabilitySpotMarket.vault,
+					assetSpotMarketVault: assetSpotMarket.vault,
+					assetTokenAccount: assetTokenAccount,
+					liabilityTokenAccount: liabilityTokenAccount,
+					tokenProgram: assetTokenProgram,
+					driftSigner: this.getStateAccount().signer,
+					instructions: anchor.web3.SYSVAR_INSTRUCTIONS_PUBKEY,
+				},
+				remainingAccounts,
+			}
+		);
+
+		return { beginSwapIx, endSwapIx };
 	}
 
 	public async liquidateBorrowForPerpPnl(
@@ -8935,6 +9199,35 @@ export class DriftClient {
 		);
 
 		return ix;
+	}
+
+	public async getPauseSpotMarketDepositWithdrawIx(
+		spotMarketIndex: number
+	): Promise<TransactionInstruction> {
+		const spotMarket = await this.getSpotMarketAccount(spotMarketIndex);
+		return this.program.instruction.pauseSpotMarketDepositWithdraw({
+			accounts: {
+				state: await this.getStatePublicKey(),
+				keeper: this.wallet.publicKey,
+				spotMarket: spotMarket.pubkey,
+				spotMarketVault: spotMarket.vault,
+			},
+		});
+	}
+
+	public async pauseSpotMarketDepositWithdraw(
+		spotMarketIndex: number,
+		txParams?: TxParams
+	): Promise<TransactionSignature> {
+		const { txSig } = await this.sendTransaction(
+			await this.buildTransaction(
+				await this.getPauseSpotMarketDepositWithdrawIx(spotMarketIndex),
+				txParams
+			),
+			[],
+			this.opts
+		);
+		return txSig;
 	}
 
 	private handleSignedTransaction(signedTxs: SignedTxData[]) {