+++ CHANGELOG.md
@@ -9,7 +9,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: update pyth lazer verification ([#1441](https://github.com/drift-labs/protocol-v2/pull/1441))
+
 ### Fixes
+
+### Breaking
+
+## [2.107.0] - 2025-01-20
+
+### Features
+
+- program: apply 10bps to protected market maker limit orders ([#1417](https://github.com/drift-labs/protocol-v2/pull/1417))
+- program: allow lighthouse at end of swap ([#1429](https://github.com/drift-labs/protocol-v2/pull/1429))
+
+### Fixes
+
+### Breaking
+
+## [2.106.0] - 2025-01-08
+
+### Features
+
+- program: liquidate spot with swap ([#1402](https://github.com/drift-labs/protocol-v2/pull/1402))
+
+### Fixes
+
+- program: account for fuel in swaps ([#1411](https://github.com/drift-labs/protocol-v2/pull/1411))
+- program: account for fuel when there is full withdraw ([#1413](https://github.com/drift-labs/protocol-v2/pull/1413))
+
+### Breaking
+
+## [2.105.0] - 2025-01-02
+
+### Features
+
+- program: add ix to pause deposits/withdraws if vault invariant broken ([#1387](https://github.com/drift-labs/protocol-v2/pull/1387))
+
+### Fixes
+
+- program: fix spot swap fuel bonus ([#1411](https://github.com/drift-labs/protocol-v2/pull/1411))
+- program: skip liq perp oracle twap check if market is in settlement ([#1406](https://github.com/drift-labs/protocol-v2/pull/1406))
 
 ### Breaking